add autentication

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\Admin\HomeController;
 use App\Http\Controllers\ThemeController;
 use Illuminate\Support\Facades\Route;
 
@@ -15,43 +16,43 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::get('/', function () {
-    return view('welcome');
+    return view('welcome1');
 });
 
 // index
-ROUTE::get('/themes','ThemeController@index')->name('themes.index');
+ROUTE::get('/themes','ThemeController@index')->name('admin.themes.index');
 
 
 
 // create
-ROUTE::get('/themes/create','ThemeController@create')->name('themes.create');
+ROUTE::get('/themes/create','ThemeController@create')->name('admin.themes.create');
 
 //store
-ROUTE::post('/themes','ThemeController@store')->name('themes.store');
+ROUTE::post('/themes','ThemeController@store')->name('admin.themes.store');
 
 // show
-ROUTE::get('/themes/{theme}','ThemeController@show')->name('themes.show');
+ROUTE::get('/themes/{theme}','ThemeController@show')->name('admin.themes.show');
 
 
 
 
 //edit
-ROUTE::get('/themes/{theme}/edit','ThemeController@edit')->name('themes.edit');
+ROUTE::get('/themes/{theme}/edit','ThemeController@edit')->name('admin.themes.edit');
 
 //update
-ROUTE::match(['patch', 'put'],'/themes/{theme}','ThemeController@update')->name('themes.update');
+ROUTE::match(['patch', 'put'],'/themes/{theme}','ThemeController@update')->name('admin.themes.update');
 
 
 //destory
-ROUTE::delete('/themes/{theme}/','ThemeController@destroy')->name('themes.destroy');
+ROUTE::delete('/themes/{theme}/','ThemeController@destroy')->name('admin.themes.destroy');
 
 
 
 
 
-// Auth::routes();
+Auth::routes();
 
-// Route::get('/home', 'HomeController@index')->name('home');
+Route::get('/home', 'HomeController@index')->name('home');
 
 
 // middleware
